% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateSummary.R
\name{calculateSummary}
\alias{calculateSummary}
\title{Calculate the summary statistics for each}
\usage{
calculateSummary(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A list containing the values of mean, standard deviation, median,
minimum, and maximum.
}
\description{
Calculate the summary statistics for each
}
