% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineCount.R
\name{centerprint}
\alias{centerprint}
\title{Prints the string in the center within the width value}
\usage{
centerprint(x, ..., width = 10)
}
\arguments{
\item{x}{A string}

\item{width}{A length of string}
}
\description{
Prints the string in the center within the width value
}
