% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INRAETableOne.R
\name{createSummary}
\alias{createSummary}
\title{Create the summary table}
\usage{
createSummary(
  x,
  y,
  data,
  max.x.level = 5,
  show.missing = TRUE,
  paired = FALSE,
  show.total = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{data}{A \code{data.frame}}

\item{paired}{If \code{paired} is TRUE, then perfrom paired t-test. The default value is \code{FALSE},}

\item{verbose}{Print the log message. The default value is \code{FALSE}}
}
\value{
A list containing the class of variable, total count of data, and
p-value. In addition, the min, max, median, sd, mean will be produced only
for continuous variable
}
\description{
It produces the list containing the detailed summary table for
the type of \code{y} (dependent variable) whether it's continuous or categorical
}
