% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INRAETableOne.R
\name{createTable}
\alias{createTable}
\title{Title}
\usage{
createTable(
  x,
  y,
  data,
  show.total = FALSE,
  paired = FALSE,
  missing = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{verbose}{}
}
\value{

}
\description{
Title
}
