% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INRAETableOne.R
\name{INRAETableOne.formula}
\alias{INRAETableOne.formula}
\title{INRAETableOne S3 Method for formula}
\usage{
\method{INRAETableOne}{formula}(formula, data, paired = FALSE, ...)
}
\arguments{
\item{formula}{A formula format}

\item{data}{A \code{data.frame}}
}
\description{
INRAETableOne S3 Method for formula
}
