% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INRAETableOne.R
\name{INRAETableOne}
\alias{INRAETableOne}
\alias{INRAETableOne.formula}
\title{Produces the descriptive summary table}
\usage{
INRAETableOne(x, ...)

\method{INRAETableOne}{formula}(
  formula,
  data,
  max.x.level = 5,
  show.missing = TRUE,
  paired = FALSE,
  show.total = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{x}{A formula indicating the dependent variable(s) on the left hand side, and
the independent variable(s) on the right hand side.}

\item{...}{More arguments can be added.}

\item{formula}{A formula format}

\item{data}{A \code{data.frame}}
}
\description{
It produces the baseline characteristics table on the dependent variables.
The maximum combination of dependent variables are two, and can be formulated with
'+' sign such as variable1 + variable2 ~ ..
Also it shows the p-value according to the type of independent variable whether
perform \code{t.test}, \code{wilcox.test}, \code{kruskal.test}, \code{anova} if continuous,
 or \code{chisq.test}, \code{fisher.test} if categorical.
}
\section{Methods (by class)}{
\itemize{
\item \code{formula}: The \code{formula} interface.
}}

\examples{
\dontrun{
library(INRAETableOne)
data(iris)
INRAETableOne(Species ~ ., iris)
}

}
