% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INRAETableOneMain.R
\name{INRAETableOneMain}
\alias{INRAETableOneMain}
\title{INRAETableOne Main function}
\usage{
INRAETableOneMain(
  formula,
  data,
  max.x.level = 5,
  show.missing = FALSE,
  paired = FALSE,
  show.total = FALSE,
  show.detail = FALSE,
  verbose = FALSE,
  origData
)
}
\description{
INRAETableOne Main function
}
