% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INRAETableOneMore.R
\name{INRAETableOneMore}
\alias{INRAETableOneMore}
\title{Generate the descriptive statistics table}
\usage{
INRAETableOneMore(
  formula,
  data,
  max.x.level = 5,
  show.missing = FALSE,
  paired = FALSE,
  show.total = FALSE,
  show.detail = FALSE,
  verbose = FALSE,
  digits = 1,
  origData = NULL
)
}
\arguments{
\item{max.x.level}{A maximum level of x level. If the level of x level is less than
max.x.level, then it consider the column as categorical.  The default value is 5}

\item{show.missing}{A Boolean expression (T/F) whether to show the missing values on the
table.}

\item{paired}{A Boolean expression (T/F) whether to perform a paired test in order to
calculate p-value.  However, the numbers of sample size at before and after
needs to be equal, otherwise, it would perform just two sample t-test.}

\item{show.total}{A Boolean expression (T/F) whether to show a total group of value.}

\item{show.detail}{A Boolean expression (T/F) whether to display the other extra
summary statistics value such as min, max, and median.}

\item{verbose}{A Boolean expression (T/F) whether to print the log messages for
every step.}

\item{x}{A formula indicating the dependent variable(s) on the left hand side, and
the independent variable(s) on the right hand side of '~'.}

\item{...}{further arguments to be passed to or from methods.}
}
\value{
An object of class "cbind.INRAETableOne".
}
\description{
The dependent and independent variable should be indicated as \code{formula} format.
The formula may contain a dot (".") to refer to "all variables in \code{data}
other than those that appear elsewhere in the formula".
The maximum combination of dependent variables are two, and can be formulated with
'+' sign such as 'dependent1' + 'dependent2' ~ 'independent1' + 'independent2' + ....
}
\details{
It produces a nicely formatted table of descriptive statistics for numeric or categorical variables
}
