% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTableOne.R
\name{makeTableOne}
\alias{makeTableOne}
\title{Generate a \code{data.frame}}
\usage{
makeTableOne(obj, digits = 1)
}
\arguments{
\item{obj}{An object returned by \code{\link{createSummary}}.}

\item{digits}{A number indicating the decimals digits}
}
\description{
Generate a \code{data.frame}
}
