% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform.chisq.test.R
\name{perform.chisq.test}
\alias{perform.chisq.test}
\title{Perform Chisq Test or Fisher's Exat Test}
\usage{
perform.chisq.test(x, y, paired = FALSE, verbose = FALSE)
}
\arguments{
\item{x}{A categorical vector}

\item{y}{A Categorical vector}

\item{paired}{If TRUE, then perform Mcnemar Test.}
}
\value{
Return the p value for categorical contingency table
}
\description{
This function calculates the p-value for the
categorical variable according to the different tpye of data.
}
