% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform.t.test.R
\name{perform.t.test}
\alias{perform.t.test}
\title{Perform t-test}
\usage{
perform.t.test(x, y, paired = FALSE, verbose = FALSE)
}
\arguments{
\item{x}{A numeric value}

\item{y}{A vector}

\item{paired}{Boolean whether perform paired t-test or not}

\item{verbose}{True as default}
}
\value{
A p-value vector according to the different type of data.
}
\description{
This function calculates the t-test according to the
different tpye of data.
}
