% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INRAETableOne.R
\name{print.cbind.INRAETableOne}
\alias{print.cbind.INRAETableOne}
\title{Print function for class of cbind.INRAETableOne}
\usage{
\method{print}{cbind.INRAETableOne}(x, ...)
}
\arguments{
\item{x}{An object of class "cbind.INRAETableOne".  It is returned from
\code{\link{cbind.INRAETableOne}}}

\item{...}{further arguments to be passed to or from methods.}
}
\description{
Print function for class of cbind.INRAETableOne
}
