% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INRAETableOne.R
\name{INRAETableOne}
\alias{INRAETableOne}
\title{It creates the descriptive summary table for}
\usage{
INRAETableOne(x, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
It creates the descriptive summary table for
}
\examples{

}
